+++ src/mefit/Firmware/EFI/Types/Enums.cs
@@ -1,7 +1,7 @@
 ﻿// Mac EFI Toolkit
 // https://github.com/MuertoGB/MacEfiToolkit
 
-// Types.cs
+// Enums.cs
 // Released under the GNU GLP v3.0
 
 namespace Mac_EFI_Toolkit.Firmware.EFI
