+++ src/mefit/Tools/Types/Structs.cs
@@ -4,9 +4,9 @@
 // Binary.cs
 // Released under the GNU GLP v3.0
 
-namespace Mac_EFI_Toolkit.Tools.Structs
+namespace Mac_EFI_Toolkit.Tools
 {
-    internal struct Binary
+    internal struct FileInfoStore
     {
         internal string FileNameExt { get; set; }
         internal string FileName { get; set; }
